selective bill works. ToDo: adding constant goods

+++ src/main/java/com/forall/laundry/model/Item.java
@@ -29,7 +29,7 @@ public class Item implements Serializable {
     @Column(unique=true,updatable=false,insertable=false,nullable=false)
     @Id
     @GeneratedValue(strategy=GenerationType.AUTO)
-    private Long item_id;
+    private Integer item_id;
     
     @Column(nullable=false)
     @Basic(fetch=FetchType.LAZY)
@@ -74,7 +74,7 @@ public class Item implements Serializable {
         this.amount = amount;
     }
    
-    public Long getItem_id() {
+    public Integer getItem_id() {
         return this.item_id;
     }
